Create abstract class for setup campaign notes
