+++ src/Notes/SetupCampaign.php
@@ -4,10 +4,6 @@ declare( strict_types=1 );
 namespace Automattic\WooCommerce\GoogleListingsAndAds\Notes;
 
 use Automattic\WooCommerce\Admin\Notes\Note as NoteEntry;
-use Automattic\WooCommerce\GoogleListingsAndAds\Ads\AdsAwareInterface;
-use Automattic\WooCommerce\GoogleListingsAndAds\Ads\AdsAwareTrait;
-use Automattic\WooCommerce\GoogleListingsAndAds\HelperTraits\Utilities;
-use Automattic\WooCommerce\GoogleListingsAndAds\PluginHelper;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -16,11 +12,7 @@ defined( 'ABSPATH' ) || exit;
  *
  * @package Automattic\WooCommerce\GoogleListingsAndAds\Notes
  */
-class SetupCampaign extends AbstractNote implements AdsAwareInterface {
-
-	use AdsAwareTrait;
-	use PluginHelper;
-	use Utilities;
+class SetupCampaign extends AbstractSetupCampaign {
 
 	/**
 	 * Get the note's unique name.
@@ -38,44 +30,17 @@ class SetupCampaign extends AbstractNote implements AdsAwareInterface {
 		$note = new NoteEntry();
 		$note->set_title( __( 'Create your first campaign to boost sales', 'google-listings-and-ads' ) );
 		$note->set_content( __( 'Leverage the power of paid ads to list products on Google Search, Shopping, YouTube, Gmail and the Display Network and drive sales.', 'google-listings-and-ads' ) );
-		$note->set_content_data( (object) [] );
-		$note->set_type( NoteEntry::E_WC_ADMIN_NOTE_INFORMATIONAL );
-		$note->set_layout( 'plain' );
-		$note->set_image( '' );
-		$note->set_name( $this->get_name() );
-		$note->set_source( $this->get_slug() );
-		$note->add_action(
-			'setup-campaign',
-			__( 'Get started', 'google-listings-and-ads' ),
-			$this->get_setup_ads_url()
-		);
+		$this->add_common_note_settings( $note );
 
 		return $note;
 	}
 
 	/**
-	 * Checks if a note can and should be added.
+	 * Get the number of days after which to add the note.
 	 *
-	 * Check if ads setup IS NOT complete
-	 * Check if it is > 3 days ago from DATE OF SETUP COMPLETION
-	 * Send notification
-	 *
-	 * @return bool
+	 * @return int
 	 */
-	public function should_be_added(): bool {
-		if ( $this->has_been_added() ) {
-			return false;
-		}
-
-		if ( $this->ads_service->is_setup_complete() ) {
-			return false;
-		}
-
-		if ( ! $this->gla_setup_for( 3 * DAY_IN_SECONDS ) ) {
-			return false;
-		}
-
-		return true;
+	protected function get_gla_setup_days(): int {
+		return 3;
 	}
-
 }
